+++ cookbooks/chub_rundeck/recipes/mysql_slave.rb
@@ -18,6 +18,8 @@
 #
 include_recipe "chub_rundeck::mysql_base"
 
+gem_package "mysql"
+
 ruby_block "start_replication" do
   block do
     require 'mysql'
@@ -56,5 +58,8 @@ ruby_block "start_replication" do
   end
 end
 
-
-
+# Sensu 
+sensu_client node.name do
+    address node[:ipaddress]
+    subscriptions node[:roles] + ["rundeck_mysql_slave"]
+end