+++ cookbooks/chub_rundeck/recipes/server.rb
@@ -39,8 +39,8 @@ end
 node[:chub_rundeck][:resources].each do |project, nodes|
   directory "#{node[:chub_rundeck][:root_project_path]}/#{project}/etc/" do
     recursive    true
-    owner     node[:chub_rundeck][:rundeck_user]
-    group     node[:chub_rundeck][:rundeck_group]
+    owner        node[:chub_rundeck][:rundeck_user]
+    group        node[:chub_rundeck][:rundeck_group]
     action       :create 
   end
 
@@ -55,6 +55,13 @@ node[:chub_rundeck][:resources].each do |project, nodes|
   end
 end
 
+rundeck_user 'api' do
+  password      'apipassword'
+  encryption    'md5'
+  roles         %w{ user admin architect deploy build api }
+  action        :create
+end
+
 cookbook_file "/etc/krb5.conf" do
   source    "krb5.conf"
   owner     "root"
@@ -62,7 +69,7 @@ cookbook_file "/etc/krb5.conf" do
 end
 
 rundeck_plugin 'rundeck-winrm-plugin-1.1.jar' do
-  checksum    "bfada6ae4215d9d44a4d1b5728bd3dde"
+  checksum    node[:chub_rundeck][:winrm_plugin_checksum]
   url         node[:chub_rundeck][:winrm_plugin_url]
   notifies    :restart, "service[rundeckd]", :delayed
 end
@@ -75,11 +82,18 @@ link "/var/lib/rundeck/bootstrap/mysql-connector-java-5.1.31-bin.jar" do
   only_if    "test -e /usr/share/java/mysql.jar"
 end
 
+dbmasters = search(:node, "mysql_master:true AND mysql_cluster_name:#{node[:mysql][:cluster_name]}")
+dbmasters = dbmasters.first
+
+
 template "#{node[:chub_rundeck][:rundeck_config_path]}/rundeck-config.properties" do
   source      'rundeck-config.properties.erb'
   owner       node[:chub_rundeck][:rundeck_user]
   group       node[:chub_rundeck][:rundeck_group]
   mode        00644
+  variables({
+       :rundeck_db_master => dbmasters.name
+  })
   notifies    :restart, 'service[rundeckd]'
 end
 
@@ -113,3 +127,12 @@ link "/etc/nginx/sites-enabled/000-default" do
 end
 
 include_recipe "rundeck::proxy"
+
+
+# Sensu 
+include_recipe "chub_sensu::client"
+
+remote_file "#{node[:chub_sensu][:root_plugin_path]}/check-http.rb" do
+    source "#{node[:chub_sensu][:root_sensu_community_plugins_repo_url]}/plugins/http/check-http.rb"
+    mode 0755
+end