+++ environments/mpqa2.rb
@@ -2,7 +2,7 @@ name "mpqa2"
 description "marketplace qa environment 2"
 cookbook "base", "= 0.1.16"
 cookbook "chub-klerk", "= 2.0.0"
-cookbook "chub-buyspace", "= 0.5.0"
+cookbook "chub-buyspace", "= 0.6.0"
 default_attributes "chub-klerk" => {
     "mongo_addresses" => [ "mpqa2db1:27017", "mpqa2db2:27017", "mpqa2db3:27017" ],
     "quartz_database_url" => "jdbc:jtds:sqlserver://sqlvm81/mpqa2_klerk_quartz;user=devusr01;password=usrdev",
@@ -116,7 +116,15 @@ default_attributes "chub-klerk" => {
     "pendingDir" => "/var/buyspace/images/forSale",
     "inProgressDir" => "/var/buyspace/images/forSale/processing",
     "doneDir" => "/var/buyspace/images/forSale/archived"
-  }  
+  },
+  "log4j" => {
+    "root_loglevel" => 'info',
+    "redis_server_host" => "mplogs01.nexus.commercehub.com"
+  },
+  "environment_level" => "testing",
+  "files" => {
+      "dailyBillingEventLocation" => '\\\\\\\\imagefs01\\\\datatransfer\\\\productstream\\\\billing'
+  },
 },
 'tomcat' => {
   'base_version' => 7,