+++ roles/rundeck_db_slave.rb
@@ -1,6 +1,11 @@
 name "rundeck_db_slave"
 description "role for rundeck_db_slave machines"
 default_attributes(
+  "sensu" => {
+      "rabbitmq" => {
+          "host" => "sensu01.nexus.commercehub.com"
+      }
+  },
   "mysql" => {
      "server_root_password" => "idontknowwhatiamdoing",
      "server_repl_password" => "idontknowwhatiamdoing",
@@ -8,4 +13,4 @@ default_attributes(
      "server_id" => 2,
   },
 )
-run_list "recipe[base::users]","recipe[chub_rundeck::mysql_slave]"
+run_list "recipe[base::users]","recipe[chub_rundeck::mysql_slave]","recipe[chub_rundeck::mysql_backup]"